+++ lib/open_weather/models/forecast/thirty_day.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require_relative 'thirty_day/thirty_day'
+require_relative 'thirty_day/forecast'
+require_relative 'thirty_day/temp'