+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.5.1 (Next)
 
+* [#42](https://github.com/dblock/open-weather-ruby-client/pull/42): Added support for the thirty day forecast in the pro 2.5 API - [@troya2](https://github.com/troya2).
 * Your contribution here.
 
 ### 0.5.0 (2024/07/03)